# docs/delivery/openapi.yaml
openapi: 3.1.0
info:
  title: MITAKO 客服 Agent 主站
  description: 提供客服前台、VIP客服工作台与运营后台服务
  version: 0.1.0
paths:
...
  /api/v1/admin/business-rules:
    get:
      tags:
      - business-rule-governance
      summary: List Rules
      operationId: list_rules_api_v1_admin_business_rules_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/admin/business-rules/{prompt_key}/versions:
    get:
      tags:
      - business-rule-governance
      summary: Get Versions
      operationId: get_versions_api_v1_admin_business_rules__prompt_key__versions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: prompt_key
        in: path
        required: true
        schema:
          type: string
          title: Prompt Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - business-rule-governance
      summary: Publish
      operationId: publish_api_v1_admin_business_rules__prompt_key__versions_post
      security:
      - HTTPBearer: []
      parameters:
      - name: prompt_key
        in: path
        required: true
        schema:
          type: string
          title: Prompt Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/business-rules/{prompt_key}/rollback:
    post:
      tags:
      - business-rule-governance
      summary: Rollback
      operationId: rollback_api_v1_admin_business_rules__prompt_key__rollback_post
      security:
      - HTTPBearer: []
      parameters:
      - name: prompt_key
        in: path
        required: true
        schema:
          type: string
          title: Prompt Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PublishBody:
      properties:
        mode:
          type: string
          title: Mode
        content:
          type: string
          title: Content
        reason:
          type: string
          title: Reason
        expected_active_version:
          type: integer
          title: Expected Active Version
      type: object
      required:
      - mode
      - content
      - reason
      - expected_active_version
      title: PublishBody
...
    RollbackBody:
      properties:
        target_version:
          type: integer
          title: Target Version
        reason:
          type: string
          title: Reason
        expected_active_version:
          type: integer
          title: Expected Active Version
      type: object
      required:
      - target_version
      - reason
      - expected_active_version
      title: RollbackBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer